ce_cloud_models/PaddleHub/CLAS/linux/requirements.txt
colorama
colorlog
easydict
filelock
flask >= 1.1.0
gitpython
numpy
matplotlib
opencv-python
packaging
paddle2onnx >= 0.5.1
paddlenlp >= 2.0.0rc5
Pillow
pyyaml
pyzmq
rarfile
tqdm
visualdl >= 2.0.0
# gunicorn not support windows
gunicorn >= 19.10.0; sys_platform != "win32"

ujson
opencv-python==4.1.2.30
pillow
tqdm
PyYAML
visualdl >= 2.0.0b
scipy
scikit-learn==0.23.2
gast==0.3.3
